pyproject.toml
[tool.poetry]
name = "magazin_restapi"
version = "0.1.0"
description = ""
authors = ["Alex <ivanovivan159753@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^22.6.2"
gino = "^1.0.1"
alembic = "^1.8.1"
asyncpg = "^0.26.0"
psycopg2-binary = "^2.9.3"
pydantic = "^1.9.2"
sanic-pydantic = "^1.3.1"
pycryptodome = "^3.15.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
sanic-testing = "^22.6.0"
pytest-asyncio = "^0.19.0"
sanic-openapi = "^21.12.0"
pytest-lazy-fixture = "^0.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
